api_server/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.4.0
torchvision==0.5.0
opencv-python==4.2.0.32
scikit-image>=0.14.2
scipy>=1.4.1
Pillow>=7.0.0
nltk==3.4.5
natsort==6.0.0
lmdb==0.97
Flask==1.1.1
gunicorn>=20.0.4